requirements.txt
streamlit==1.24.0
SpeechRecognition==3.10.0
pydub==0.25.1
fpdf2==2.7.4
python-dotenv==1.0.0
